fixup! add get-deps script
